refactor(types): split communication.ts into entities/{channel,message,board-message,note,poll,poll-vote,acknowledgement}.ts
