+++ packages/types/src/__tests__/encryption.test.ts
@@ -1,15 +1,6 @@
 import { describe, expectTypeOf, it } from "vitest";
 
 import type { AuditActor, AuditLogEntry } from "../audit-log.js";
-import type {
-  AcknowledgementRequest,
-  BoardMessage,
-  Channel,
-  ChatMessage,
-  Note,
-  Poll,
-  PollVote,
-} from "../communication.js";
 import type { KdfMasterKey } from "../crypto-keys.js";
 import type {
   BucketEncrypted,
@@ -73,11 +64,18 @@ import type {
   T1EncryptedBlob,
   T2EncryptedBlob,
 } from "../encryption.js";
+import type { AcknowledgementRequest } from "../entities/acknowledgement.js";
+import type { BoardMessage } from "../entities/board-message.js";
+import type { Channel } from "../entities/channel.js";
 import type { CustomFront } from "../entities/custom-front.js";
 import type { FrontingComment } from "../entities/fronting-comment.js";
 import type { FrontingSession } from "../entities/fronting-session.js";
 import type { MemberPhoto } from "../entities/member-photo.js";
 import type { Member } from "../entities/member.js";
+import type { ChatMessage } from "../entities/message.js";
+import type { Note } from "../entities/note.js";
+import type { PollVote } from "../entities/poll-vote.js";
+import type { Poll } from "../entities/poll.js";
 import type { Group } from "../groups.js";
 import type {
   BucketId,
@@ -92,7 +90,7 @@ import type {
 } from "../ids.js";
 import type { JournalEntry, WikiPage } from "../journal.js";
 import type { LifecycleEvent } from "../lifecycle.js";
-import type { Relationship } from "../structure.js";
+import type { Relationship } from "../entities/relationship.js";
 import type { TimerConfig } from "../timer.js";
 import type { UnixMillis } from "../timestamps.js";
 import type { Serialize } from "../type-assertions.js";
